Don't insert previously negated items in `Applications.append_if_new()`
We add logic in `Applications.append_if_new()` to only insert
non-negated items if they aren't already on our internal negation list.

Additionally, we drop items from the negation list, once we've applied
the negation once either in `append_if_new()` or in `merge_unique()` so
that patterns like adding, then removing and then adding an item again
have the expected result of the item being present in the final list.

This fixes the issue where it wasn't possible to preemptively remove an
entry from the applications list in a multi-dimensional hierarchy, e.g.
in a [Commodore] global defaults repository where we may want to exclude
applications for a certain Kubernetes distribution regardless of the
cloud on which a cluster with that distribution is running.

[Commodore]: https://syn.tools/commodore